+++ subsystems/routine/__init__.py
@@ -1,0 +1,3 @@
+from subsystems.routine.subsystem import RoutineSubsystem
+
+__all__ = ["RoutineSubsystem"]